Add basic layout to dashboard
