--- src/components/Dashboard/Header.js
@@ -1,1 +1,0 @@
-
