+++ src/styles/Dashboard.css
@@ -22,12 +22,45 @@
   --mainGold: #daa520;
 }
 
- 
- .sidebar {
+#header {
+  display: flex;
+  justify-content: center;
+  align-items: center;
   left: 0;
+  right: 0;
+  top: 0;
+  position: fixed;
+  height: 90px;
+  padding-right: 1rem;
+  padding-left: 1rem;
+}
+
+ .hide-sidebar,
+ .show-sidebar {
+  left: 0;
+  top: 0;
   width: 15%;
-  top: 87.16px;
   position: fixed;
   height: 100vh;
+  transition: 0.3s ease;
   background: var(--darkBlue);
+  padding-left: 1rem;
+  z-index: 999;
+ }
+
+ .hide-sidebar {
+  transform: translate(-100%);
+ }
+
+ .show-sidebar {
+  transform: translate(0);
+ }
+
+ .logo-wrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  height: 90px;
+  margin-bottom: 1rem;
  }
